Variable naming and default messages
Renamed a few variables for consistency in the input functions (the construct - test - facet triplet is now in all the functions for easier understanding).
Also I added messages for automatically generated default values. If such a value has to be generated by default, a message in the console informs the user to which value it has been set (e.g. facet circle size).